Added smaller bug fixes. Fixed Small Ghast could not fly. #21 Fixed Snails spawn on the sea and stay on top. #23 Fixed Server crasher like #24

+++ src/main/java/de/markusbordihn/playercompanions/data/PlayerCompanionsServerData.java
@@ -77,7 +77,7 @@ public class PlayerCompanionsServerData extends SavedData {
 
     if (Boolean.TRUE.equals(COMMON.dataBackupEnabled.get())) {
       nextBackupTime =
-          (int) java.time.Instant.now().getEpochSecond() + (60 * COMMON.dataBackupInterval.get());
+          java.time.Instant.now().getEpochSecond() + (60 * COMMON.dataBackupInterval.get());
       log.info("Enable automatic data backups every {} minutes, next backup will run at {} ...",
           COMMON.dataBackupInterval.get(), nextBackupTime);
     } else {
